[code_cleanup-1] Clean up using IntelliJ's 'Silent Code Cleanup'

+++ core/src/net/sf/openrocket/masscalc/MassCalculation.java
@@ -106,7 +106,7 @@ public class MassCalculation {
 	
 	@Override
 	public int hashCode() {
-		return (int) (this.centerOfMass.hashCode());
+		return this.centerOfMass.hashCode();
 	}
 
 	public MassCalculation( final Type _type, final FlightConfiguration _config, final double _time,
